joiner/requirements.txt
telethon==1.34.0
aiofiles==23.2.1
python-dotenv==1.0.0
loguru==0.7.2
psutil==5.9.6
requests==2.31.0
beautifulsoup4==4.12.2
